relocate files:
* sample_correlation_matrices.hpp -> sampling/

* test.hpp -> test/sampling_correlation_matrices_test.cpp

* ReHMC.cpp -> random_walks/gaussian_ReHMC_correlation.hpp
